advanced/mcp_adk/pyproject.toml
[project]
name = "mcp-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.1",
    "google-adk>=1.19.0",
    "pydantic>=2.12.4",
    "rich>=14.2.0",
]
